refactor: restructure execution code and enhance testing framework
- Consolidate and refactor execution logic in `execution.go` and `execution_test.go`
- Remove deprecated integration tests and related files
- Update transaction handling to use byte arrays instead of custom types
- Improve error handling and logging in the execution client methods
- Adjust Go module dependencies and update to Go 1.24.0
- Remove unnecessary build commands from the Makefile
- Enhance test setup for better isolation and cleanup
