+++ Makefile
@@ -13,13 +13,6 @@ help: Makefile
 	@sed -n 's/^##//p' $< | column -t -s ':' |  sed -e 's/^/ /'
 .PHONY: help
 
-## build: build evm-middleware binary
-build: build/evm-middleware
-.PHONY: build
-
-build/evm-middleware: cmd/evm-middleware/main.go execution.go go.mod go.sum
-	@echo "Building build/evm-middleware"
-	@go build -o build/evm-middleware ./cmd/evm-middleware
 
 ## clean: clean testcache
 clean:
